// src/modules/product-adm/facade/product-adm.facade.interface.ts
export interface AddProductFacadeInputDto {
    id?: string;
    name: string;
    description: string;
    purchasePrice: number;
    stock: number;
  }
  
  export interface CheckStockFacadeInputDto {
    productId: string;
  }
  
  export interface CheckStockFacadeOutputDto {
    productId: string;
    stock: number;
  }
  
  export default interface ProductAdmFacadeInterface {
    addProduct(input: AddProductFacadeInputDto): Promise<void>;
    checkStock(
      input: CheckStockFacadeInputDto
    ): Promise<CheckStockFacadeOutputDto>;
  }